+++ i6_models/assemblies/conformer/__init__.py
@@ -1,0 +1,1 @@
+from .conformer import *
